plant is food +minor change
-plant herite IFood
-zone ne prennes plus en compte les object qui doivent disparaitre
-debugg comportementPlant

+++ ecosysteme/ecosysteme/Models/ComportementPlantDefault.cs
@@ -19,7 +19,7 @@ namespace ecosysteme.Models
         public void UpdateEtat(Plant thisObject)
         {
             (int,int) energie = thisObject.getEnergie();
-            if (energie.Item1 / energie.Item2 < 0.5 && thisObject.CanEat())
+            if ((double)energie.Item1 / energie.Item2 < 0.5 && thisObject.CanEat())
             {
                 etat = etatComportementBase.Alimentation;
             }
